--- demo_client/public/manifest.json
@@ -1,9 +1,0 @@
-{
-  "short_name": "Emotion",
-  "name": "Inference with Transformers in the Browser",
-  "icons": [],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
