--- demo_client/public/xtremedistil-int8.onnx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:63b7f456c4fee56059785088f97c7a160c6dd8a439f63285501ec9a4c993f9f9
-size 13088747
