--- demo_client/public/xtremedistill-go-emotion-int8.onnx
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:7acc0eb39fce713acfd9a8a1a655b2ccd9c660e4acb20058cbc6e5fb1e32b16f
-size 22878749
