--- demo_client/src/game_process.js
@@ -1,3 +1,0 @@
-/**
- * 
- */
