--- demo_client/src/setupProxy.js
@@ -1,7 +1,0 @@
-module.exports = function (app) {
-    app.use(function (req, res, next) {
-        res.setHeader("Cross-Origin-Opener-Policy", "same-origin");
-        res.setHeader("Cross-Origin-Embedder-Policy", "require-corp");
-        next();
-    });
-};
